fix: ngx.null userdata and entity checks
Previously, ngx.null as a 'null' string caused all sorts of breakage.
ngx.null has now been replaced with a userdata to match the behavior of
OpenResty.
To match ngx.null with lua-cjson, the json library has been patched to
ignore *every* userdata. This is not an issue for the schema library.
Finally, process_auto_field is now executed as an "insert"
unconditionally to popualte missing fields.