fix: make sure 'enabled' and 'sandbox_enabled' are set to use the sandbox (#16)
Since goks only cares about the plugin validation part of the
Kong Lua tree, it's run in a way that's not mounting the default
configuration file. The sandbox mode is triggered via some
config flags, which were missing (and therefore unset).

This adds a patch change to make sure the sandbox is enabled.